some adjustments to Space, + beast skeletons
made some adjustments to the way things are stored spatially, they're more clear and easy now, and they work with texture Y-sorting.
Also added some skeletons of beast logic. It's basically going to need to pick a position to move towards and a speed each frame, based on a state machine.